Fix overflow.

+++ source/thread_scheduler.c
@@ -102,7 +102,7 @@ static void s_thread_fn(void *arg) {
         int64_t timeout = 0;
         if (next_scheduled_task == UINT64_MAX) {
             /* at least wake up once per 30 seconds. */
-            timeout = 30 * AWS_TIMESTAMP_NANOS;
+            timeout = (int64_t)30 * (int64_t)AWS_TIMESTAMP_NANOS;
         } else {
             timeout = (int64_t)(next_scheduled_task - current_time);
         }
